requirements.txt
paho-mqtt
numpy
opencv-python
scikit-image
matplotlib
pandas
statsmodels
tk
Pillow
pexpect